Port from ppm42ms branch

+++ src/pulses_arm.cpp
@@ -159,8 +159,8 @@ void setupPulsesPPM()                   // Don't enable interrupts through here
   else
     pwmptr->PWM_CH_NUM[3].PWM_CMR &= ~0x00000200 ;  // CPOL
 
-  uint16_t rest = 22500u * 2; //Minimum Framelen=22.5 ms
-  rest += (int16_t(g_model.ppmFrameLength)) * 1000;
+  uint32_t rest = 22500u * 2; //Minimum Framelen=22.5 ms
+  rest += (int32_t(g_model.ppmFrameLength)) * 1000;
   for (uint32_t i = 0; i < p; i++) { //NUM_CHNOUT
     int16_t v = limit((int16_t)-PPM_range, g_chans512[i], (int16_t)PPM_range) + 2*PPM_CENTER;
     rest -= (v);
